Safari-611.1.10/LayoutTests/js/regexp-extended-characters-match-expected.txt

This test checks regular expressions using extended (> 255) characters and character classes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS (new RegExp("[Ā-ā]")).exec("a") is null
PASS (new RegExp("[Ā]")).exec("a") is null
PASS (new RegExp("Ā")).exec("a") is null
PASS (new RegExp("[a]")).exec("a").toString() is "a"
PASS (new RegExp("[Ā-āa]")).exec("a").toString() is "a"
PASS (new RegExp("[Āa]")).exec("a").toString() is "a"
PASS (new RegExp("a")).exec("a").toString() is "a"
PASS (new RegExp("[a-Ā]")).exec("a").toString() is "a"
PASS (new RegExp("[Ā]")).exec("Ā").toString() is "Ā"
PASS (new RegExp("[Ā-ā]")).exec("Ā").toString() is "Ā"
PASS (new RegExp("Ā")).exec("Ā").toString() is "Ā"
PASS successfullyParsed is true

TEST COMPLETE
